Made big components smaller
