+++ client/src/Components/GameMap.tsx
@@ -1,0 +1,32 @@
+export function GameMap() {
+  return (
+    <div>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={4 * 150}
+        viewBox="0 -0.5 150 184"
+        shapeRendering="crispEdges">
+        <metadata>
+          Made with Pixels to Svg https://codepen.io/shshaw/pen/XbxvNj
+        </metadata>
+        <path
+          stroke="#595652"
+          d="M0 0h35M60 0h15M86 0h16M129 0h21M0 1h36M59 1h16M86 1h17M128 1h22M0 2h37M58 2h17M76 2h1M84 2h1M86 2h18M127 2h23M0 3h38M57 3h18M86 3h19M126 3h24M0 4h39M56 4h19M76 4h1M84 4h1M86 4h20M125 4h25M0 5h77M84 5h66M0 6h77M84 6h66M0 7h77M84 7h66M0 8h78M83 8h67M0 9h79M82 9h68M0 10h80M81 10h69M0 11h150M0 12h150M0 13h150M0 14h150M0 15h150M0 16h150M0 17h150M0 18h150M0 19h150M0 20h150M0 21h150M0 22h150M0 23h150M0 24h150M0 25h150M0 26h150M0 27h150M0 28h150M0 29h150M0 30h150M0 31h150M0 32h150M0 33h150M0 34h4M5 34h6M12 34h7M20 34h6M27 34h105M133 34h6M140 34h10M0 35h4M6 35h5M12 35h7M21 35h5M27 35h105M134 35h5M140 35h10M0 36h4M6 36h4M13 36h6M21 36h4M28 36h104M134 36h4M141 36h9M0 37h4M7 37h3M13 37h6M22 37h3M28 37h104M135 37h3M141 37h9M0 38h4M7 38h3M14 38h5M22 38h3M29 38h103M135 38h3M142 38h8M0 39h4M7 39h3M14 39h5M22 39h3M29 39h103M135 39h3M142 39h8M0 40h3M7 40h3M14 40h4M22 40h3M29 40h102M135 40h3M142 40h8M0 41h3M8 41h2M14 41h4M23 41h2M29 41h102M136 41h2M142 41h8M0 42h2M8 42h1M14 42h3M23 42h1M29 42h101M136 42h1M142 42h8M0 43h2M8 43h1M14 43h3M23 43h1M29 43h101M136 43h1M142 43h8M0 44h2M14 44h3M29 44h101M142 44h8M0 45h1M15 45h1M30 45h99M143 45h7M0 46h1M31 46h98M144 46h6M0 47h1M31 47h98M144 47h6M0 48h150M0 49h150M0 50h52M53 50h6M60 50h8M69 50h6M76 50h24M101 50h6M108 50h42M0 51h52M54 51h5M60 51h8M70 51h5M76 51h24M102 51h5M108 51h42M0 52h52M54 52h4M61 52h7M70 52h4M77 52h23M102 52h4M109 52h41M0 53h52M55 53h3M61 53h7M71 53h3M77 53h23M103 53h3M109 53h41M0 54h52M55 54h3M62 54h6M71 54h3M78 54h22M103 54h3M110 54h40M0 55h52M55 55h3M62 55h6M71 55h3M78 55h22M103 55h3M110 55h40M0 56h51M55 56h3M62 56h5M71 56h3M78 56h21M103 56h3M110 56h40M0 57h51M56 57h2M62 57h5M72 57h2M78 57h21M104 57h2M110 57h40M0 58h50M56 58h1M62 58h4M72 58h1M78 58h20M104 58h1M110 58h40M0 59h50M56 59h1M62 59h4M72 59h1M78 59h20M104 59h1M110 59h40M0 60h50M62 60h4M78 60h20M110 60h40M0 61h49M63 61h2M79 61h18M111 61h39M0 62h49M64 62h1M80 62h17M112 62h38M0 63h49M64 63h1M80 63h17M112 63h38M0 64h150M0 65h150M0 66h19M20 66h6M27 66h41M69 66h6M76 66h40M117 66h6M124 66h26M0 67h19M21 67h5M27 67h41M70 67h5M76 67h40M118 67h5M124 67h26M0 68h19M21 68h4M28 68h40M70 68h4M77 68h39M118 68h4M125 68h25M0 69h19M22 69h3M28 69h40M71 69h3M77 69h39M119 69h3M125 69h25M0 70h19M22 70h3M29 70h39M71 70h3M78 70h38M119 70h3M126 70h24M0 71h19M22 71h3M29 71h39M71 71h3M78 71h38M119 71h3M126 71h24M0 72h18M22 72h3M29 72h38M71 72h3M78 72h37M119 72h3M126 72h24M0 73h18M23 73h2M29 73h38M72 73h2M78 73h37M120 73h2M126 73h24M0 74h17M23 74h1M29 74h37M72 74h1M78 74h36M120 74h1M126 74h24M0 75h17M23 75h1M29 75h37M72 75h1M78 75h36M120 75h1M126 75h24M0 76h17M29 76h37M78 76h36M126 76h24M0 77h16M30 77h35M79 77h34M127 77h23M0 78h16M31 78h34M80 78h33M128 78h22M0 79h16M31 79h34M80 79h33M128 79h22M0 80h150M0 81h150M0 82h3M4 82h6M11 82h8M20 82h6M27 82h9M37 82h6M44 82h56M101 82h6M108 82h42M0 83h3M5 83h5M11 83h8M21 83h5M27 83h9M38 83h5M44 83h56M102 83h5M108 83h42M0 84h3M5 84h4M12 84h7M21 84h4M28 84h8M38 84h4M45 84h55M102 84h4M109 84h41M0 85h3M6 85h3M12 85h7M22 85h3M28 85h8M39 85h3M45 85h55M103 85h3M109 85h41M0 86h3M6 86h3M13 86h6M22 86h3M29 86h7M39 86h3M46 86h54M103 86h3M110 86h40M0 87h3M6 87h3M13 87h6M22 87h3M29 87h7M39 87h3M46 87h54M103 87h3M110 87h40M0 88h2M6 88h3M13 88h5M22 88h3M29 88h6M39 88h3M46 88h53M103 88h3M110 88h40M0 89h2M7 89h2M13 89h5M23 89h2M29 89h6M40 89h2M46 89h53M104 89h2M110 89h40M0 90h1M7 90h1M13 90h4M23 90h1M29 90h5M40 90h1M46 90h52M104 90h1M110 90h40M0 91h1M7 91h1M13 91h4M23 91h1M29 91h5M40 91h1M46 91h52M104 91h1M110 91h40M0 92h1M13 92h4M29 92h5M46 92h52M110 92h40M14 93h2M30 93h3M47 93h50M111 93h39M15 94h1M31 94h2M48 94h49M112 94h38M15 95h1M31 95h2M48 95h49M112 95h38M0 96h150M0 97h150M0 98h67M68 98h6M75 98h9M85 98h6M92 98h8M101 98h6M108 98h23M132 98h6M139 98h11M0 99h67M69 99h5M75 99h9M86 99h5M92 99h8M102 99h5M108 99h23M133 99h5M139 99h11M0 100h67M69 100h4M76 100h8M86 100h4M93 100h7M102 100h4M109 100h22M133 100h4M140 100h10M0 101h67M70 101h3M76 101h8M87 101h3M93 101h7M103 101h3M109 101h22M134 101h3M140 101h10M0 102h67M70 102h3M77 102h7M87 102h3M94 102h6M103 102h3M110 102h21M134 102h3M141 102h9M0 103h67M70 103h3M77 103h7M87 103h3M94 103h6M103 103h3M110 103h21M134 103h3M141 103h9M0 104h66M70 104h3M77 104h6M87 104h3M94 104h5M103 104h3M110 104h20M134 104h3M141 104h9M0 105h66M71 105h2M77 105h6M88 105h2M94 105h5M104 105h2M110 105h20M135 105h2M141 105h9M0 106h65M71 106h1M77 106h5M88 106h1M94 106h4M104 106h1M110 106h19M135 106h1M141 106h9M0 107h65M71 107h1M77 107h5M88 107h1M94 107h4M104 107h1M110 107h19M135 107h1M141 107h9M0 108h65M77 108h5M94 108h4M110 108h19M141 108h9M0 109h64M78 109h3M95 109h2M111 109h17M142 109h8M0 110h64M79 110h2M96 110h1M112 110h16M143 110h7M0 111h64M79 111h2M96 111h1M112 111h16M143 111h7M0 112h150M0 113h150M0 114h3M4 114h6M11 114h139M0 115h3M5 115h5M11 115h139M0 116h3M5 116h4M12 116h138M0 117h3M6 117h3M12 117h138M0 118h3M6 118h3M13 118h137M0 119h3M6 119h3M13 119h137M0 120h2M6 120h3M13 120h137M0 121h2M7 121h2M13 121h137M0 122h1M7 122h1M13 122h137M0 123h1M7 123h1M13 123h137M0 124h1M13 124h137M14 125h136M15 126h135M15 127h135M0 128h150M0 129h150M0 130h36M37 130h6M44 130h40M85 130h6M92 130h24M117 130h6M124 130h26M0 131h36M38 131h5M44 131h40M86 131h5M92 131h24M118 131h5M124 131h26M0 132h36M38 132h4M45 132h39M86 132h4M93 132h23M118 132h4M125 132h25M0 133h36M39 133h3M45 133h39M87 133h3M93 133h23M119 133h3M125 133h25M0 134h36M39 134h3M46 134h38M87 134h3M94 134h22M119 134h3M126 134h24M0 135h36M39 135h3M46 135h38M87 135h3M94 135h22M119 135h3M126 135h24M0 136h35M39 136h3M46 136h37M87 136h3M94 136h21M119 136h3M126 136h24M0 137h35M40 137h2M46 137h37M88 137h2M94 137h21M120 137h2M126 137h24M0 138h34M40 138h1M46 138h36M88 138h1M94 138h20M120 138h1M126 138h24M0 139h34M40 139h1M46 139h36M88 139h1M94 139h20M120 139h1M126 139h24M0 140h34M46 140h36M94 140h20M126 140h24M0 141h33M47 141h34M95 141h18M127 141h23M0 142h33M48 142h33M96 142h17M128 142h22M0 143h33M48 143h33M96 143h17M128 143h22M0 144h150M0 145h150M0 146h20M21 146h6M28 146h72M101 146h6M108 146h42M0 147h20M22 147h5M28 147h72M102 147h5M108 147h42M0 148h20M22 148h4M29 148h71M102 148h4M109 148h41M0 149h20M23 149h3M29 149h71M103 149h3M109 149h41M0 150h20M23 150h3M30 150h70M103 150h3M110 150h40M0 151h20M23 151h3M30 151h70M103 151h3M110 151h40M0 152h19M23 152h3M30 152h69M103 152h3M110 152h40M0 153h19M24 153h2M30 153h69M104 153h2M110 153h40M0 154h18M24 154h1M30 154h68M104 154h1M110 154h40M0 155h18M24 155h1M30 155h68M104 155h1M110 155h40M0 156h18M30 156h68M110 156h40M0 157h17M31 157h66M111 157h39M0 158h17M32 158h65M112 158h38M0 159h17M32 159h65M112 159h38M0 160h150M0 161h150M0 162h36M37 162h6M44 162h24M69 162h6M76 162h8M85 162h6M92 162h40M133 162h6M140 162h10M0 163h36M38 163h5M44 163h24M70 163h5M76 163h8M86 163h5M92 163h40M134 163h5M140 163h10M0 164h36M38 164h4M45 164h23M70 164h4M77 164h7M86 164h4M93 164h39M134 164h4M141 164h9M0 165h36M39 165h3M45 165h23M71 165h3M77 165h7M87 165h3M93 165h39M135 165h3M141 165h9M0 166h36M39 166h3M46 166h22M71 166h3M78 166h6M87 166h3M94 166h38M135 166h3M142 166h8M0 167h36M39 167h3M46 167h22M71 167h3M78 167h6M87 167h3M94 167h38M135 167h3M142 167h8M0 168h35M39 168h3M46 168h21M71 168h3M78 168h5M87 168h3M94 168h37M135 168h3M142 168h8M0 169h35M40 169h2M46 169h21M72 169h2M78 169h5M88 169h2M94 169h37M136 169h2M142 169h8M0 170h34M40 170h1M46 170h20M72 170h1M78 170h4M88 170h1M94 170h36M136 170h1M142 170h8M0 171h34M40 171h1M46 171h20M72 171h1M78 171h4M88 171h1M94 171h36M136 171h1M142 171h8M0 172h34M46 172h20M78 172h4M94 172h36M142 172h8M0 173h33M47 173h18M79 173h2M95 173h34M143 173h7M0 174h33M48 174h17M80 174h1M96 174h33M144 174h6M0 175h33M48 175h17M80 175h1M96 175h33M144 175h6M0 176h150M0 177h150M0 178h150M0 179h150M0 180h150M0 181h150M0 182h150M0 183h150"
+        />
+        <path
+          stroke="#4da681"
+          d="M35 0h25M75 0h11M102 0h27M36 1h23M76 1h9M103 1h25M37 2h21M77 2h7M104 2h23M38 3h19M77 3h1M79 3h3M83 3h1M105 3h21M39 4h17M77 4h2M80 4h1M82 4h2M106 4h19M77 5h2M82 5h2M77 6h7M77 7h7M78 8h5M79 9h3M80 10h1"
+        />
+        <path
+          stroke="#000000"
+          d="M75 1h1M85 1h1M75 2h1M85 2h1M75 3h2M78 3h1M82 3h1M84 3h2M75 4h1M79 4h1M81 4h1M85 4h1M80 5h1M4 34h1M11 34h1M19 34h1M26 34h1M132 34h1M139 34h1M4 35h2M11 35h1M19 35h2M26 35h1M132 35h2M139 35h1M4 36h2M10 36h1M12 36h1M19 36h2M25 36h1M27 36h1M132 36h2M138 36h1M140 36h1M4 37h1M6 37h1M10 37h1M12 37h1M19 37h1M21 37h1M25 37h1M27 37h1M132 37h1M134 37h1M138 37h1M140 37h1M4 38h1M6 38h1M10 38h1M13 38h1M19 38h1M21 38h1M25 38h1M28 38h1M132 38h1M134 38h1M138 38h1M141 38h1M4 39h1M6 39h1M10 39h1M13 39h1M19 39h1M21 39h1M25 39h1M28 39h1M132 39h1M134 39h1M138 39h1M141 39h1M3 40h1M6 40h1M10 40h1M13 40h1M18 40h1M21 40h1M25 40h1M28 40h1M131 40h1M134 40h1M138 40h1M141 40h1M3 41h1M7 41h1M10 41h1M13 41h1M18 41h1M22 41h1M25 41h1M28 41h1M131 41h1M135 41h1M138 41h1M141 41h1M2 42h1M7 42h1M9 42h1M13 42h1M17 42h1M22 42h1M24 42h1M28 42h1M130 42h1M135 42h1M137 42h1M141 42h1M2 43h1M9 43h1M13 43h1M17 43h1M24 43h1M28 43h1M130 43h1M137 43h1M141 43h1M2 44h1M13 44h1M17 44h1M28 44h1M130 44h1M141 44h1M1 45h1M14 45h1M16 45h1M29 45h1M129 45h1M142 45h1M15 46h1M30 46h1M143 46h1M15 47h1M30 47h1M143 47h1M52 50h1M59 50h1M68 50h1M75 50h1M100 50h1M107 50h1M52 51h2M59 51h1M68 51h2M75 51h1M100 51h2M107 51h1M52 52h2M58 52h1M60 52h1M68 52h2M74 52h1M76 52h1M100 52h2M106 52h1M108 52h1M52 53h1M54 53h1M58 53h1M60 53h1M68 53h1M70 53h1M74 53h1M76 53h1M100 53h1M102 53h1M106 53h1M108 53h1M52 54h1M54 54h1M58 54h1M61 54h1M68 54h1M70 54h1M74 54h1M77 54h1M100 54h1M102 54h1M106 54h1M109 54h1M52 55h1M54 55h1M58 55h1M61 55h1M68 55h1M70 55h1M74 55h1M77 55h1M100 55h1M102 55h1M106 55h1M109 55h1M51 56h1M54 56h1M58 56h1M61 56h1M67 56h1M70 56h1M74 56h1M77 56h1M99 56h1M102 56h1M106 56h1M109 56h1M51 57h1M55 57h1M58 57h1M61 57h1M67 57h1M71 57h1M74 57h1M77 57h1M99 57h1M103 57h1M106 57h1M109 57h1M50 58h1M55 58h1M57 58h1M61 58h1M66 58h1M71 58h1M73 58h1M77 58h1M98 58h1M103 58h1M105 58h1M109 58h1M50 59h1M57 59h1M61 59h1M66 59h1M73 59h1M77 59h1M98 59h1M105 59h1M109 59h1M50 60h1M61 60h1M66 60h1M77 60h1M98 60h1M109 60h1M49 61h1M62 61h1M65 61h1M78 61h1M97 61h1M110 61h1M63 62h1M79 62h1M111 62h1M63 63h1M79 63h1M111 63h1M19 66h1M26 66h1M68 66h1M75 66h1M116 66h1M123 66h1M19 67h2M26 67h1M68 67h2M75 67h1M116 67h2M123 67h1M19 68h2M25 68h1M27 68h1M68 68h2M74 68h1M76 68h1M116 68h2M122 68h1M124 68h1M19 69h1M21 69h1M25 69h1M27 69h1M68 69h1M70 69h1M74 69h1M76 69h1M116 69h1M118 69h1M122 69h1M124 69h1M19 70h1M21 70h1M25 70h1M28 70h1M68 70h1M70 70h1M74 70h1M77 70h1M116 70h1M118 70h1M122 70h1M125 70h1M19 71h1M21 71h1M25 71h1M28 71h1M68 71h1M70 71h1M74 71h1M77 71h1M116 71h1M118 71h1M122 71h1M125 71h1M18 72h1M21 72h1M25 72h1M28 72h1M67 72h1M70 72h1M74 72h1M77 72h1M115 72h1M118 72h1M122 72h1M125 72h1M18 73h1M22 73h1M25 73h1M28 73h1M67 73h1M71 73h1M74 73h1M77 73h1M115 73h1M119 73h1M122 73h1M125 73h1M17 74h1M22 74h1M24 74h1M28 74h1M66 74h1M71 74h1M73 74h1M77 74h1M114 74h1M119 74h1M121 74h1M125 74h1M17 75h1M24 75h1M28 75h1M66 75h1M73 75h1M77 75h1M114 75h1M121 75h1M125 75h1M17 76h1M28 76h1M66 76h1M77 76h1M114 76h1M125 76h1M16 77h1M29 77h1M65 77h1M78 77h1M113 77h1M126 77h1M30 78h1M79 78h1M127 78h1M30 79h1M79 79h1M127 79h1M3 82h1M10 82h1M19 82h1M26 82h1M36 82h1M43 82h1M100 82h1M107 82h1M3 83h2M10 83h1M19 83h2M26 83h1M36 83h2M43 83h1M100 83h2M107 83h1M3 84h2M9 84h1M11 84h1M19 84h2M25 84h1M27 84h1M36 84h2M42 84h1M44 84h1M100 84h2M106 84h1M108 84h1M3 85h1M5 85h1M9 85h1M11 85h1M19 85h1M21 85h1M25 85h1M27 85h1M36 85h1M38 85h1M42 85h1M44 85h1M100 85h1M102 85h1M106 85h1M108 85h1M3 86h1M5 86h1M9 86h1M12 86h1M19 86h1M21 86h1M25 86h1M28 86h1M36 86h1M38 86h1M42 86h1M45 86h1M100 86h1M102 86h1M106 86h1M109 86h1M3 87h1M5 87h1M9 87h1M12 87h1M19 87h1M21 87h1M25 87h1M28 87h1M36 87h1M38 87h1M42 87h1M45 87h1M100 87h1M102 87h1M106 87h1M109 87h1M2 88h1M5 88h1M9 88h1M12 88h1M18 88h1M21 88h1M25 88h1M28 88h1M35 88h1M38 88h1M42 88h1M45 88h1M99 88h1M102 88h1M106 88h1M109 88h1M2 89h1M6 89h1M9 89h1M12 89h1M18 89h1M22 89h1M25 89h1M28 89h1M35 89h1M39 89h1M42 89h1M45 89h1M99 89h1M103 89h1M106 89h1M109 89h1M1 90h1M6 90h1M8 90h1M12 90h1M17 90h1M22 90h1M24 90h1M28 90h1M34 90h1M39 90h1M41 90h1M45 90h1M98 90h1M103 90h1M105 90h1M109 90h1M1 91h1M8 91h1M12 91h1M17 91h1M24 91h1M28 91h1M34 91h1M41 91h1M45 91h1M98 91h1M105 91h1M109 91h1M1 92h1M12 92h1M17 92h1M28 92h1M34 92h1M45 92h1M98 92h1M109 92h1M0 93h1M13 93h1M16 93h1M29 93h1M33 93h1M46 93h1M97 93h1M110 93h1M14 94h1M30 94h1M47 94h1M111 94h1M14 95h1M30 95h1M47 95h1M111 95h1M67 98h1M74 98h1M84 98h1M91 98h1M100 98h1M107 98h1M131 98h1M138 98h1M67 99h2M74 99h1M84 99h2M91 99h1M100 99h2M107 99h1M131 99h2M138 99h1M67 100h2M73 100h1M75 100h1M84 100h2M90 100h1M92 100h1M100 100h2M106 100h1M108 100h1M131 100h2M137 100h1M139 100h1M67 101h1M69 101h1M73 101h1M75 101h1M84 101h1M86 101h1M90 101h1M92 101h1M100 101h1M102 101h1M106 101h1M108 101h1M131 101h1M133 101h1M137 101h1M139 101h1M67 102h1M69 102h1M73 102h1M76 102h1M84 102h1M86 102h1M90 102h1M93 102h1M100 102h1M102 102h1M106 102h1M109 102h1M131 102h1M133 102h1M137 102h1M140 102h1M67 103h1M69 103h1M73 103h1M76 103h1M84 103h1M86 103h1M90 103h1M93 103h1M100 103h1M102 103h1M106 103h1M109 103h1M131 103h1M133 103h1M137 103h1M140 103h1M66 104h1M69 104h1M73 104h1M76 104h1M83 104h1M86 104h1M90 104h1M93 104h1M99 104h1M102 104h1M106 104h1M109 104h1M130 104h1M133 104h1M137 104h1M140 104h1M66 105h1M70 105h1M73 105h1M76 105h1M83 105h1M87 105h1M90 105h1M93 105h1M99 105h1M103 105h1M106 105h1M109 105h1M130 105h1M134 105h1M137 105h1M140 105h1M65 106h1M70 106h1M72 106h1M76 106h1M82 106h1M87 106h1M89 106h1M93 106h1M98 106h1M103 106h1M105 106h1M109 106h1M129 106h1M134 106h1M136 106h1M140 106h1M65 107h1M72 107h1M76 107h1M82 107h1M89 107h1M93 107h1M98 107h1M105 107h1M109 107h1M129 107h1M136 107h1M140 107h1M65 108h1M76 108h1M82 108h1M93 108h1M98 108h1M109 108h1M129 108h1M140 108h1M64 109h1M77 109h1M81 109h1M94 109h1M97 109h1M110 109h1M128 109h1M141 109h1M78 110h1M95 110h1M111 110h1M142 110h1M78 111h1M95 111h1M111 111h1M142 111h1M3 114h1M10 114h1M3 115h2M10 115h1M3 116h2M9 116h1M11 116h1M3 117h1M5 117h1M9 117h1M11 117h1M3 118h1M5 118h1M9 118h1M12 118h1M3 119h1M5 119h1M9 119h1M12 119h1M2 120h1M5 120h1M9 120h1M12 120h1M2 121h1M6 121h1M9 121h1M12 121h1M1 122h1M6 122h1M8 122h1M12 122h1M1 123h1M8 123h1M12 123h1M1 124h1M12 124h1M0 125h1M13 125h1M14 126h1M14 127h1M36 130h1M43 130h1M84 130h1M91 130h1M116 130h1M123 130h1M36 131h2M43 131h1M84 131h2M91 131h1M116 131h2M123 131h1M36 132h2M42 132h1M44 132h1M84 132h2M90 132h1M92 132h1M116 132h2M122 132h1M124 132h1M36 133h1M38 133h1M42 133h1M44 133h1M84 133h1M86 133h1M90 133h1M92 133h1M116 133h1M118 133h1M122 133h1M124 133h1M36 134h1M38 134h1M42 134h1M45 134h1M84 134h1M86 134h1M90 134h1M93 134h1M116 134h1M118 134h1M122 134h1M125 134h1M36 135h1M38 135h1M42 135h1M45 135h1M84 135h1M86 135h1M90 135h1M93 135h1M116 135h1M118 135h1M122 135h1M125 135h1M35 136h1M38 136h1M42 136h1M45 136h1M83 136h1M86 136h1M90 136h1M93 136h1M115 136h1M118 136h1M122 136h1M125 136h1M35 137h1M39 137h1M42 137h1M45 137h1M83 137h1M87 137h1M90 137h1M93 137h1M115 137h1M119 137h1M122 137h1M125 137h1M34 138h1M39 138h1M41 138h1M45 138h1M82 138h1M87 138h1M89 138h1M93 138h1M114 138h1M119 138h1M121 138h1M125 138h1M34 139h1M41 139h1M45 139h1M82 139h1M89 139h1M93 139h1M114 139h1M121 139h1M125 139h1M34 140h1M45 140h1M82 140h1M93 140h1M114 140h1M125 140h1M33 141h1M46 141h1M81 141h1M94 141h1M113 141h1M126 141h1M47 142h1M95 142h1M127 142h1M47 143h1M95 143h1M127 143h1M20 146h1M27 146h1M100 146h1M107 146h1M20 147h2M27 147h1M100 147h2M107 147h1M20 148h2M26 148h1M28 148h1M100 148h2M106 148h1M108 148h1M20 149h1M22 149h1M26 149h1M28 149h1M100 149h1M102 149h1M106 149h1M108 149h1M20 150h1M22 150h1M26 150h1M29 150h1M100 150h1M102 150h1M106 150h1M109 150h1M20 151h1M22 151h1M26 151h1M29 151h1M100 151h1M102 151h1M106 151h1M109 151h1M19 152h1M22 152h1M26 152h1M29 152h1M99 152h1M102 152h1M106 152h1M109 152h1M19 153h1M23 153h1M26 153h1M29 153h1M99 153h1M103 153h1M106 153h1M109 153h1M18 154h1M23 154h1M25 154h1M29 154h1M98 154h1M103 154h1M105 154h1M109 154h1M18 155h1M25 155h1M29 155h1M98 155h1M105 155h1M109 155h1M18 156h1M29 156h1M98 156h1M109 156h1M17 157h1M30 157h1M97 157h1M110 157h1M31 158h1M111 158h1M31 159h1M111 159h1M36 162h1M43 162h1M68 162h1M75 162h1M84 162h1M91 162h1M132 162h1M139 162h1M36 163h2M43 163h1M68 163h2M75 163h1M84 163h2M91 163h1M132 163h2M139 163h1M36 164h2M42 164h1M44 164h1M68 164h2M74 164h1M76 164h1M84 164h2M90 164h1M92 164h1M132 164h2M138 164h1M140 164h1M36 165h1M38 165h1M42 165h1M44 165h1M68 165h1M70 165h1M74 165h1M76 165h1M84 165h1M86 165h1M90 165h1M92 165h1M132 165h1M134 165h1M138 165h1M140 165h1M36 166h1M38 166h1M42 166h1M45 166h1M68 166h1M70 166h1M74 166h1M77 166h1M84 166h1M86 166h1M90 166h1M93 166h1M132 166h1M134 166h1M138 166h1M141 166h1M36 167h1M38 167h1M42 167h1M45 167h1M68 167h1M70 167h1M74 167h1M77 167h1M84 167h1M86 167h1M90 167h1M93 167h1M132 167h1M134 167h1M138 167h1M141 167h1M35 168h1M38 168h1M42 168h1M45 168h1M67 168h1M70 168h1M74 168h1M77 168h1M83 168h1M86 168h1M90 168h1M93 168h1M131 168h1M134 168h1M138 168h1M141 168h1M35 169h1M39 169h1M42 169h1M45 169h1M67 169h1M71 169h1M74 169h1M77 169h1M83 169h1M87 169h1M90 169h1M93 169h1M131 169h1M135 169h1M138 169h1M141 169h1M34 170h1M39 170h1M41 170h1M45 170h1M66 170h1M71 170h1M73 170h1M77 170h1M82 170h1M87 170h1M89 170h1M93 170h1M130 170h1M135 170h1M137 170h1M141 170h1M34 171h1M41 171h1M45 171h1M66 171h1M73 171h1M77 171h1M82 171h1M89 171h1M93 171h1M130 171h1M137 171h1M141 171h1M34 172h1M45 172h1M66 172h1M77 172h1M82 172h1M93 172h1M130 172h1M141 172h1M33 173h1M46 173h1M65 173h1M78 173h1M81 173h1M94 173h1M129 173h1M142 173h1M47 174h1M79 174h1M95 174h1M143 174h1M47 175h1M79 175h1M95 175h1M143 175h1"
+        />
+        <path stroke="#ac3232" d="M79 5h1M81 5h1" />
+        <path
+          stroke="#323c39"
+          d="M11 36h1M26 36h1M139 36h1M5 37h1M11 37h1M20 37h1M26 37h1M133 37h1M139 37h1M5 38h1M11 38h2M20 38h1M26 38h2M133 38h1M139 38h2M5 39h1M11 39h2M20 39h1M26 39h2M133 39h1M139 39h2M4 40h2M11 40h2M19 40h2M26 40h2M132 40h2M139 40h2M4 41h3M11 41h2M19 41h3M26 41h2M132 41h3M139 41h2M3 42h4M10 42h3M18 42h4M25 42h3M131 42h4M138 42h3M3 43h5M10 43h3M18 43h5M25 43h3M131 43h5M138 43h3M3 44h10M18 44h10M131 44h10M2 45h12M17 45h12M130 45h12M1 46h14M16 46h14M129 46h14M1 47h14M16 47h14M129 47h14M59 52h1M75 52h1M107 52h1M53 53h1M59 53h1M69 53h1M75 53h1M101 53h1M107 53h1M53 54h1M59 54h2M69 54h1M75 54h2M101 54h1M107 54h2M53 55h1M59 55h2M69 55h1M75 55h2M101 55h1M107 55h2M52 56h2M59 56h2M68 56h2M75 56h2M100 56h2M107 56h2M52 57h3M59 57h2M68 57h3M75 57h2M100 57h3M107 57h2M51 58h4M58 58h3M67 58h4M74 58h3M99 58h4M106 58h3M51 59h5M58 59h3M67 59h5M74 59h3M99 59h5M106 59h3M51 60h10M67 60h10M99 60h10M50 61h12M66 61h12M98 61h12M49 62h14M65 62h14M97 62h14M49 63h14M65 63h14M97 63h14M26 68h1M75 68h1M123 68h1M20 69h1M26 69h1M69 69h1M75 69h1M117 69h1M123 69h1M20 70h1M26 70h2M69 70h1M75 70h2M117 70h1M123 70h2M20 71h1M26 71h2M69 71h1M75 71h2M117 71h1M123 71h2M19 72h2M26 72h2M68 72h2M75 72h2M116 72h2M123 72h2M19 73h3M26 73h2M68 73h3M75 73h2M116 73h3M123 73h2M18 74h4M25 74h3M67 74h4M74 74h3M115 74h4M122 74h3M18 75h5M25 75h3M67 75h5M74 75h3M115 75h5M122 75h3M18 76h10M67 76h10M115 76h10M17 77h12M66 77h12M114 77h12M16 78h14M65 78h14M113 78h14M16 79h14M65 79h14M113 79h14M10 84h1M26 84h1M43 84h1M107 84h1M4 85h1M10 85h1M20 85h1M26 85h1M37 85h1M43 85h1M101 85h1M107 85h1M4 86h1M10 86h2M20 86h1M26 86h2M37 86h1M43 86h2M101 86h1M107 86h2M4 87h1M10 87h2M20 87h1M26 87h2M37 87h1M43 87h2M101 87h1M107 87h2M3 88h2M10 88h2M19 88h2M26 88h2M36 88h2M43 88h2M100 88h2M107 88h2M3 89h3M10 89h2M19 89h3M26 89h2M36 89h3M43 89h2M100 89h3M107 89h2M2 90h4M9 90h3M18 90h4M25 90h3M35 90h4M42 90h3M99 90h4M106 90h3M2 91h5M9 91h3M18 91h5M25 91h3M35 91h5M42 91h3M99 91h5M106 91h3M2 92h10M18 92h10M35 92h10M99 92h10M1 93h12M17 93h12M34 93h12M98 93h12M0 94h14M16 94h14M33 94h14M97 94h14M0 95h14M16 95h14M33 95h14M97 95h14M74 100h1M91 100h1M107 100h1M138 100h1M68 101h1M74 101h1M85 101h1M91 101h1M101 101h1M107 101h1M132 101h1M138 101h1M68 102h1M74 102h2M85 102h1M91 102h2M101 102h1M107 102h2M132 102h1M138 102h2M68 103h1M74 103h2M85 103h1M91 103h2M101 103h1M107 103h2M132 103h1M138 103h2M67 104h2M74 104h2M84 104h2M91 104h2M100 104h2M107 104h2M131 104h2M138 104h2M67 105h3M74 105h2M84 105h3M91 105h2M100 105h3M107 105h2M131 105h3M138 105h2M66 106h4M73 106h3M83 106h4M90 106h3M99 106h4M106 106h3M130 106h4M137 106h3M66 107h5M73 107h3M83 107h5M90 107h3M99 107h5M106 107h3M130 107h5M137 107h3M66 108h10M83 108h10M99 108h10M130 108h10M65 109h12M82 109h12M98 109h12M129 109h12M64 110h14M81 110h14M97 110h14M128 110h14M64 111h14M81 111h14M97 111h14M128 111h14M10 116h1M4 117h1M10 117h1M4 118h1M10 118h2M4 119h1M10 119h2M3 120h2M10 120h2M3 121h3M10 121h2M2 122h4M9 122h3M2 123h5M9 123h3M2 124h10M1 125h12M0 126h14M0 127h14M43 132h1M91 132h1M123 132h1M37 133h1M43 133h1M85 133h1M91 133h1M117 133h1M123 133h1M37 134h1M43 134h2M85 134h1M91 134h2M117 134h1M123 134h2M37 135h1M43 135h2M85 135h1M91 135h2M117 135h1M123 135h2M36 136h2M43 136h2M84 136h2M91 136h2M116 136h2M123 136h2M36 137h3M43 137h2M84 137h3M91 137h2M116 137h3M123 137h2M35 138h4M42 138h3M83 138h4M90 138h3M115 138h4M122 138h3M35 139h5M42 139h3M83 139h5M90 139h3M115 139h5M122 139h3M35 140h10M83 140h10M115 140h10M34 141h12M82 141h12M114 141h12M33 142h14M81 142h14M113 142h14M33 143h14M81 143h14M113 143h14M27 148h1M107 148h1M21 149h1M27 149h1M101 149h1M107 149h1M21 150h1M27 150h2M101 150h1M107 150h2M21 151h1M27 151h2M101 151h1M107 151h2M20 152h2M27 152h2M100 152h2M107 152h2M20 153h3M27 153h2M100 153h3M107 153h2M19 154h4M26 154h3M99 154h4M106 154h3M19 155h5M26 155h3M99 155h5M106 155h3M19 156h10M99 156h10M18 157h12M98 157h12M17 158h14M97 158h14M17 159h14M97 159h14M43 164h1M75 164h1M91 164h1M139 164h1M37 165h1M43 165h1M69 165h1M75 165h1M85 165h1M91 165h1M133 165h1M139 165h1M37 166h1M43 166h2M69 166h1M75 166h2M85 166h1M91 166h2M133 166h1M139 166h2M37 167h1M43 167h2M69 167h1M75 167h2M85 167h1M91 167h2M133 167h1M139 167h2M36 168h2M43 168h2M68 168h2M75 168h2M84 168h2M91 168h2M132 168h2M139 168h2M36 169h3M43 169h2M68 169h3M75 169h2M84 169h3M91 169h2M132 169h3M139 169h2M35 170h4M42 170h3M67 170h4M74 170h3M83 170h4M90 170h3M131 170h4M138 170h3M35 171h5M42 171h3M67 171h5M74 171h3M83 171h5M90 171h3M131 171h5M138 171h3M35 172h10M67 172h10M83 172h10M131 172h10M34 173h12M66 173h12M82 173h12M130 173h12M33 174h14M65 174h14M81 174h14M129 174h14M33 175h14M65 175h14M81 175h14M129 175h14"
+        />
+      </svg>
+    </div>
+  );
+}
