+++ client/src/Components/Pause.tsx
@@ -1,0 +1,9 @@
+export function Pause() {
+  return (
+    <div className="text-center">
+      <div className="p-3">
+        <button className="text-white hover:text-yellow-500">Pause</button>
+      </div>
+    </div>
+  );
+}
