+++ client/src/Components/Time.tsx
@@ -1,0 +1,7 @@
+export function Time() {
+  return (
+    <div className="flex justify-center text-center">
+      <div className="bg-yellow-300 p-8 drop-shadow-2xl">00:00:00</div>
+    </div>
+  );
+}
